refactor(risk): pindah risk engine ke client-side (web-app)
Hapus supabase/functions/calculate-risk/ (ga dipake lagi)
Buat web-app/src/lib/risk/ — risk engine murni client-side:
  - engine.ts: scoring functions + analyze (port dari Python + Deno)
  - elevation.ts: Open-Meteo elevation API
  - slope.ts: 4-point gradient calculation
  - weather.ts: Open-Meteo forecast API (rain, temp, humidity)
  - soil.ts: ISRIC SoilGrids API
  - index.ts: unified calculateRisk()
  - Semua pake fetch() + AbortController, cache in-memory
  - No Supabase, no edge function dependency

Update:
  - ReportFormPage.tsx: import risk types dari lib/risk/
  - types/laporan.ts: hapus RiskFactorReport (pindah ke lib/risk/)
  - Fix unclosed <div> di ReportFormPage.tsx (pre-existing bug)
